修改了gpt的接口，添加了yolo定位图片模型训练的部分，添加了机构迭代的流程

+++ request.py
@@ -14,7 +14,7 @@ from tqdm import tqdm
 paper_save_path='./paper_download/'
 if not os.path.exists(paper_save_path):
     os.mkdir(paper_save_path)
-authorization="Bearer eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ1c2VyLWNlbnRlciIsImV4cCI6MTcwNzM4NDQ0NiwiaWF0IjoxNzA3MzgzNTQ2LCJqdGkiOiJjbjI5bHVnM3IwN2FyZGdtMmFyZyIsInR5cCI6ImFjY2VzcyIsInN1YiI6ImNtbDN2bmN1ZHU2NXZtNTFkb2NnIiwic3BhY2VfaWQiOiJjbWwzdm5jdWR1NjV2bTUxZG9jMCIsImFic3RyYWN0X3VzZXJfaWQiOiJjbWwzdm5jdWR1NjV2bTUxZG9jZyJ9.-qWhqRtMMIgjXUGpCA607h2v7TSbTxYbkMyhY6sAfPwmAmRYWvpexvNnlOQQmIBCtYPd3e_Jh6oGnAnYaH04DQ"
+authorization="Bearer eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ1c2VyLWNlbnRlciIsImV4cCI6MTcwODM0NzI2MiwiaWF0IjoxNzA4MzQ2MzYyLCJqdGkiOiJjbjlrbnVtY3A3ZmQxM2k4cWQ4MCIsInR5cCI6ImFjY2VzcyIsInN1YiI6ImNtbDN2bmN1ZHU2NXZtNTFkb2NnIiwic3BhY2VfaWQiOiJjbWwzdm5jdWR1NjV2bTUxZG9jMCIsImFic3RyYWN0X3VzZXJfaWQiOiJjbWwzdm5jdWR1NjV2bTUxZG9jZyJ9.N1XnSNh5Yb_kI3zMkv7N_5ArTIGV0cwTXXZ1c4zsQV0RgGSs_mI1su4TKTQLNBayCxStig5isePcsGOqwo_OTQ"
 proxies = {
     'http': 'http://127.0.0.1:8466',
     'https': 'https://127.0.0.1:8466'
@@ -41,14 +41,15 @@ def get_mechanism(file_name):
     return get_mechanism_chatgpt(text)
 
 def get_mechanism_chatgpt(text):
-    content=f"我接下来给你一页论文，请找出其中全部的作者的学校或者研究机构，并以缩写的形式回复我,中间用, 隔开。例如：MIT, ZJU, PKU。请仅回答作者的学校或者研究机构的缩写，不要回答其他的信息。论文：```{text}```"
+    content=f"I will give you a page of paper next. Please find out the names of all the authors’ schools or research institutions, and reply to me in the form of name:name abbreviation, separated by '||' For example:“Massachusetts Institute of Technology:MIT||Peking University:PKU||Tencent AI Lab:Tencent AI Lab||Deemos Technology:Deemos”。If there are multiple identical schools or research institutions. Only answer the research school, institution and its abbreviation according to the prescribed format. Do not answer other information.paper：```{text}```"
     return chatgpt(content=content,s=s)
 
 def get_summary(arxiv_pdf_url):
     return kimi(content=f"请用300字总结，分三段 <url id=\"\" type=\"url\" status=\"\" title=\"\" wc=\"\">{arxiv_pdf_url}</url>",s=s,authorization=authorization)
 if not resume:
     for file in os.listdir(paper_save_path):
-        os.remove(paper_save_path+file)
+        if file.endswith('.pdf'):
+            os.remove(paper_save_path+file)
 with open('paper.yaml', 'r', encoding='utf-8') as f:
     topics = yaml.load(f.read(), Loader=yaml.FullLoader)
 if resume:
@@ -70,6 +71,7 @@ for Paper in pbar:
             task_content[Paper]['summary']=get_summary(Paper.replace('abs','pdf'))
     except Exception as e:
         print(e) 
+        print(f'Error in {Paper}')
         json.dump(task_content,open('result.json','w',encoding='utf-8'),ensure_ascii=False)
     json.dump(task_content,open('result.json','w',encoding='utf-8'),ensure_ascii=False)
 with open('result.md','w',encoding='utf-8') as f:
@@ -80,7 +82,8 @@ with open('result.md','w',encoding='utf-8') as f:
                 title,publish_datetime,Project_Page='','',''
             else:
                 title,publish_datetime,Project_Page=task_content[Paper]['title_datetime_url']
-            mechanism=task_content[Paper]['mechanism'] if task_content[Paper]['mechanism'] else ''
+            # mechanism=task_content[Paper]['mechanism'] if task_content[Paper]['mechanism'] else ''
+            mechanism=''
             f.write('## '+title+'\n\n')
             f.write(f'{publish_datetime}｜{mechanism}'+'\n\n')
             if Project_Page:
@@ -89,4 +92,14 @@ with open('result.md','w',encoding='utf-8') as f:
                 f.write(f'<u>{Paper}</u>'+'\n\n')
             paper_summary=task_content[Paper]['summary'] if task_content[Paper]['summary'] else ''
             f.write(paper_summary+'\n\n')
-    
+mecanism_dict={i.split(':')[0]:i.split(':')[1] for i in open('mechanism_list.txt','r',encoding='utf-8').read().split('\n') if i}
+with open('mechanism_check.txt','w',encoding='utf-8') as f:
+    for paper in task_content:
+        if task_content[paper]['mechanism']:
+            f.write(task_content[paper]['title_datetime_url'][0]+'\n')
+            for mecanism in task_content[paper]['mechanism'].replace('&nbsp;',' ').split('||'):
+                mecanism_split=mecanism.split(':')
+                mecanism_long=mecanism_split[0].strip()
+                mecanism_short=mecanism_dict.get(mecanism_long) if mecanism_dict.get(mecanism_long) else mecanism_split[1].strip()
+                f.write(f'{mecanism_long}:{mecanism_short}\n')
+            f.write('--------------------------------------------------------------------------------\n')
